Fix compile/link errors too.

+++ exec/load/sbbslist_lib.js
@@ -1,4 +1,5 @@
-// $Id$
+// $Id: sbbslist_lib.js,v 1.24 2020/04/22 00:54:58 rswindell Exp $
+// @format.tab-size 4
 
 // Synchronet BBS List (SBL) v4 Library
 
@@ -26,6 +27,7 @@ const max_len = {
 	updated_on:			28,
 	verified_on:		28,
 	service_address:	35,		// Increased from 28
+	service_description:15,
 	software:			15,
 	since:				4,		/* just the year portion of the first_online date */
 	nodes:				4,
@@ -36,6 +38,10 @@ const max_len = {
 	msgs:				8,
 	files:				7,
 	storage:			7,
+	min_rate:			5,
+	max_rate:			5,
+	port:				5,
+	protocol:			15,
 	protocols:			60,
 	tcp_services:		33,
 	udp_services:		20,
@@ -44,6 +50,29 @@ const max_len = {
 	networks:			60,
 	description:		250,
 };
+
+// Some of these max values are hard-technical limits (e.g. port), some are legacy
+// SBL limits (e.g. 16-bit min/max_rate) and some are just limited for cosmetic reasons
+const max_val = {
+	port:				65535,
+	nodes:				255,
+	users:				9999,
+	subs:				99999,
+	dirs:				99999,
+	doors:				99999,
+	msgs:				99999999,
+	files:				9999999,
+	min_rate:			65535,
+	max_rate:			65535,
+};
+
+// Services supported by BBS clients (e.g. SyncTERM)
+const common_bbs_services = [
+    "telnet",
+    "rlogin",
+	"ssh",
+	"modem"
+];
 
 function compare(a, b)
 {
@@ -85,6 +114,11 @@ function verify_compare(a, b)
     }
 
     return diff;
+}
+
+function active_compare(a, b)
+{
+	return last_active(b) - last_active(a);
 }
 
 function date_string(date)
@@ -151,41 +185,41 @@ function property_value(bbs, property)
 				value = bbs.first_online.substring(0,4);
 			break;
 		case "nodes":
-			if(bbs.terminal && bbs.terminal.nodes)
+			if(bbs.terminal && bbs.terminal.nodes !== undefined)
 				value = bbs.terminal.nodes;
 			break;
 		case "users":
-			if(bbs.total && bbs.total.users)
+			if(bbs.total && bbs.total.users !== undefined)
 				value = bbs.total.users;
 			break;
 		case "subs":
-			if(bbs.total && bbs.total.subs)
+			if(bbs.total && bbs.total.subs !== undefined)
 				value = bbs.total.subs;
 			break;
 		case "dirs":
-			if(bbs.total && bbs.total.dirs)
+			if(bbs.total && bbs.total.dirs !== undefined)
 				value = bbs.total.dirs;
 			break;
 		case "doors":
-			if(bbs.total && bbs.total.doors)
+			if(bbs.total && bbs.total.doors !== undefined)
 				value = bbs.total.doors;
 			break;
 		case "msgs":
-			if(bbs.total && bbs.total.msgs) {
+			if(bbs.total && bbs.total.msgs !== undefined) {
 				value = bbs.total.msgs;
 				if(value >= 1000000)
 					value = Math.ceil(value / 1000000) + "M";
 			}
 			break;
 		case "files":
-			if(bbs.total && bbs.total.files) {
+			if(bbs.total && bbs.total.files !== undefined) {
 				value = bbs.total.files;
 				if(value >= 1000000)
 					value = Math.ceil(value / 1000000) + "M";
 			}
 			break;
 		case "storage":
-			if(bbs.total && bbs.total.storage) {
+			if(bbs.total && bbs.total.storage && !isNaN(bbs.total.storage)) {
 				if(bbs.total.storage > 1024*1024*1024*1024)
 					value = Math.ceil(bbs.total.storage / (1024*1024*1024*1024)) + "TB";
 				else if(bbs.total.storage > 1024*1024*1024)
@@ -256,6 +290,14 @@ function property_value(bbs, property)
 	return value;
 }
 
+function numeric_sort_value(value)
+{
+	value = parseInt(value, 10);
+	if(isNaN(value))
+		return -1;
+	return value;
+}
+
 /* Some properties are nested within arrays of objects, lets simplify those properties here */
 function property_sort_value(bbs, property)
 {
@@ -268,15 +310,35 @@ function property_sort_value(bbs, property)
 			return bbs.name;
 		case "storage":
 			if(bbs.total)
-				return bbs.total.storage;
+				return numeric_sort_value(bbs.total.storage);
 			break;
 		case "files":
 			if(bbs.total)
-				return bbs.total.files;
+				return numeric_sort_value(bbs.total.files);
 			brak;
 		case "msgs":
 			if(bbs.total)
-				return bbs.total.msgs;
+				return numeric_sort_value(bbs.total.msgs);
+			break;
+		case "dirs":
+			if(bbs.total)
+				return numeric_sort_value(bbs.total.dirs);
+			break;
+		case "subs":
+			if(bbs.total)
+				return numeric_sort_value(bbs.total.subs);
+			break;
+		case "users":
+			if(bbs.total)
+				return numeric_sort_value(bbs.total.users);
+			break;
+		case "doors":
+			if(bbs.total)
+				return numeric_sort_value(bbs.total.doors);
+			break;
+		case "nodes":
+			if(bbs.terminal)
+				return numeric_sort_value(bbs.terminal.nodes);
 			break;
 		case "created_on":
 			return new Date(bbs.entry.created.on).valueOf();
@@ -296,6 +358,8 @@ function sort(list, property)
 {
 	if(property=="verify")
 		list.sort(verify_compare);
+	else if(property=="active")
+		list.sort(active_compare);
 	else {	
 		if(property)
 			sort_property=property;
@@ -339,62 +403,106 @@ function parse(f)
 	return list;
 }
 
+function lock_fname()
+{
+	return list_fname + '.lock';
+}
+
+function lock(op, timeout /* in seconds */, max)
+{
+	var start = time();
+	if(timeout === undefined)
+		timeout = 60;
+	if(max === undefined)
+		max = 24 * 60 * 60;
+
+	while(!file_mutex(lock_fname(), op, max)) {
+		if(time() - start >= timeout) {
+        		log(LOG_ERR, "SBBSLIST: Timeout locking " + list_fname + " for " + op);
+			return false;
+		}
+		sleep(1000);
+	}
+	return true;
+}
+
+function unlock()
+{
+	return file_remove(lock_fname());
+}
+
 function read_list()
 {
+	if(!lock("read"))
+		return false;
     var f = new File(list_fname);
-    log(LOG_INFO, "Opening list file: " + f.name);
+    log(LOG_DEBUG, "Opening list file: " + f.name);
     if(!f.open("r")) {
         log(LOG_ERR, "SBBSLIST: Error " + f.error + " opening " + f.name);
+		unlock();
         return [];
     }
 	var list = parse(f);
     f.close();
+	unlock();
     return list;
 }
 
 function write_list(list)
 {
+	if(!lock("write"))
+		return false;
     var out = new File(list_fname);
-    log(LOG_INFO, "SBBSLIST: Opening / creating list file: " + list_fname);
+    log(LOG_DEBUG, "SBBSLIST: Opening / creating list file: " + list_fname);
     if(!out.open("w+")) {
         log(LOG_ERR, "SBBSLIST: Error " + out.error + " creating " + out.name);
+		unlock();
         return false;
     }
-    log(LOG_INFO, "SBBSLIST: Writing list file: " + out.name + " (" + list.length + " BBS entries)");
+    log(LOG_DEBUG, "SBBSLIST: Writing list file: " + out.name + " (" + list.length + " BBS entries)");
     out.write(JSON.stringify(list, null, 4));
     out.close();
+	unlock();
     return true;
 }
 
 function append(bbs)
 {
+	if(!lock("append"))
+		return false;
     var f = new File(list_fname);
-    log(LOG_INFO, "SBBSLIST: Opening / creating list file: " + list_fname);
+    log(LOG_DEBUG, "SBBSLIST: Opening / creating list file: " + list_fname);
     if(!f.open(f.exists ? 'r+':'w+')) {
         log(LOG_ERR, "SBBSLIST: Error " + f.error + " creating " + f.name);
+		unlock();
         return false;
     }
 	var list = parse(f);
 	list.push(bbs);
-    log(LOG_INFO, "SBBSLIST: Writing list file: " + f.name + " (" + list.length + " BBS entries)");
+    log(LOG_DEBUG, "SBBSLIST: Writing list file: " + f.name + " (" + list.length + " BBS entries)");
 	f.truncate();
     f.write(JSON.stringify(list, null, 4));
     f.close();
+	unlock();
     return true;
 }
 
 function remove(bbs)
 {
+	if(!lock("remove"))
+		return false;
     var f = new File(list_fname);
     log(LOG_INFO, "SBBSLIST: Opening / creating list file: " + list_fname);
     if(!f.open('r+')) {
         log(LOG_ERR, "SBBSLIST: Error " + f.error + " opening " + f.name);
+		unlock();
         return false;
     }
     var list = parse(f);
 	var index = find_value(list, "name", bbs.name); 
 	if(index < 0 ) {
 		f.close();
+		unlock();
 		return false;
 	}
 	list.splice(index, 1);
@@ -402,21 +510,26 @@ function remove(bbs)
 	f.truncate();
     f.write(JSON.stringify(list, null, 4));
     f.close();
+	unlock();
     return true;
 }
 
 function replace(bbs)
 {
+	if(!lock("replace"))
+		return false;
     var f = new File(list_fname);
     log(LOG_INFO, "SBBSLIST: Opening / creating list file: " + list_fname);
     if(!f.open('r+')) {
         log(LOG_ERR, "SBBSLIST: Error " + f.error + " opening " + f.name);
+		unlock();
         return false;
     }
     var list = parse(f);
 	var index = find_value(list, "name", bbs.name); 
 	if(index < 0 ) {
 		f.close();
+		unlock();
 		return false;
 	}
 	list[index] = bbs;
@@ -424,10 +537,11 @@ function replace(bbs)
 	f.truncate();
     f.write(JSON.stringify(list, null, 4));
     f.close();
+	unlock();
     return true;
 }
 
-function remove_dupes(list)
+function remove_dupes(list, verbose)
 {
     var new_list=[];
     var names=[];
@@ -435,25 +549,70 @@ function remove_dupes(list)
     var i;
 
     for(i in list) {
-		var name = list[i].name.toLowerCase();
+		var bbs = list[i];
+		var name = bbs.name.toLowerCase();
 		if(name.slice(-4) == " bbs")
 			name = name.slice(0, -4);
 		if(name.slice(0, 4) == "the ")
 			name = name.slice(4);
 		if(names.indexOf(name) >= 0)
 			continue;
-		if(list[i].service.length
-			&& address.indexOf(list[i].service[0].address.toLowerCase()) >= 0)
+		if(bbs.service.length
+			&& address.indexOf(bbs.service[0].address.toLowerCase()) >= 0) {
+			if(verbose)
+				print(format("%-25s : Duplicated name", bbs.name));
 			continue;
+		}
 		names.push(name);
-		address.push(list[i].service[0].address.toLowerCase());
-		new_list.push(list[i]);
+		address.push(bbs.service[0].address.toLowerCase());
+		new_list.push(bbs);
 	}
     return new_list;
 }
 
+function last_active(bbs)
+{
+	var active = 0;
+	if(bbs.entry.created)
+		active = new Date(bbs.entry.created.on);
+	var updated = 0;
+	if(bbs.entry.updated)
+		updated = new Date(bbs.entry.updated.on);
+	if(updated > active)
+		active = updated;
+	var verified = 0;
+	if(bbs.entry.verified)
+		verified = new Date(bbs.entry.verified.on);
+	if(verified > active)
+		active = verified;
+	return active;
+}
+
+function remove_inactive(list, max, verbose)
+{
+    var new_list=[];
+    var i;
+
+    for(i in list) {
+		var bbs = list[i];
+		var active = last_active(bbs);
+		var diff = new Date().valueOf() - active.valueOf();
+		var days = diff / (24 * 60 * 60 * 1000);
+		if(days > max) {
+			if(verbose)
+				print(format("%-25s : Inactive since %.10s (%d days)"
+					, bbs.name, active.toISOString(), days));
+			continue;
+		}
+		new_list.push(list[i]);
+	}
+	return new_list;
+}
+
 function imsg_capable_system(bbs)
 {
+	if(bbs.imsg_capable == true)
+		return true;
 	if(!bbs.entry.autoverify || !bbs.entry.autoverify.last_success)
 		return false;
 	var services = bbs.entry.autoverify.last_success.other_services;
@@ -564,21 +723,73 @@ function syncterm_list(list, dir)
     }
 	f.writeln(format("; Exported from %s on %s", system.name, new Date().toString()));
 	f.writeln();
+	var sections = [];
     for(i in list) {
         for(j in list[i].service) {
             if(!list[i].service[j].protocol)
                 continue;
+			if(common_bbs_services.indexOf(list[i].service[j].protocol.toLowerCase()) < 0)
+				continue;
+			var section;
             if(j > 0)
-                f.writeln(format("[%-23.23s %6.6s]", list[i].name, list[i].service[j].protocol));
+                section = format("%-23.23s %6.6s", list[i].name, list[i].service[j].protocol);
             else
-                f.writeln(format("[%s]", list[i].name));
-            f.writeln(format("ConnectionType=%s", list[i].service[j].protocol));
-            f.writeln(format("Address=%s", list[i].service[j].address));
+                section = list[i].name;
+			if(sections.indexOf(section) >= 0) {	// duplicate
+				if(list[i].service[j].description)
+					section = format("%-20.20s%10.10s", list[i].name, list[i].service[j].description);
+				else if(list[i].service[j].port)
+					section = format("%-24.24s %5u", list[i].name, list[i].service[j].port);
+				if(sections.indexOf(section) >= 0)	// duplicate
+					continue;
+			}
+			sections.push(section);
+			sections.push(format("%-23.23s %6.6s", list[i].name, list[i].service[j].protocol));
+			f.writeln(format("[%s]", section));
+            f.writeln(format("\tConnectionType=%s", list[i].service[j].protocol));
+            f.writeln(format("\tAddress=%s", list[i].service[j].address));
+			if(list[i].service[j].port)
+				f.writeln(format("\tPort=%s", list[i].service[j].port));
+			if(list[i].service[j].description
+				&& (list[i].service[j].description.toUpperCase().indexOf("PETSCII") == 0
+				|| list[i].service[j].description.toUpperCase().indexOf("COMMODORE") == 0)) {
+				f.writeln(format("\tScreenMode=%s"
+					,list[i].service[j].description.indexOf("80") >= 0 ? "C128-80col" : "C64"));
+				f.writeln("\tNoStatus=true");
+			}
             f.writeln();
         }
     }
     f.close();
 	return f.name;
+}
+
+function share_list(list, qwk)
+{
+	var new_list = [];
+	for(i in list) {
+		var bbs = list[i];
+		delete bbs.entry.autoverify;
+		delete bbs.entry.exported;
+		if(qwk) {
+			if(bbs.entry.created.at && isNaN(parseInt(bbs.entry.created.at, 10)))
+				bbs.entry.created.at = system.qwk_id + "/" + bbs.entry.created.at;
+			else
+				bbs.entry.created.at = system.qwk_id;
+			if(bbs.entry.updated.at && isNaN(parseInt(bbs.entry.updated.at, 10)))
+				bbs.entry.updated.at = system.qwk_id + "/" + bbs.entry.updated.at;
+			else
+				bbs.entry.updated.at = system.qwk_id;
+		} else {
+			if(isNaN(parseInt(bbs.entry.created.at, 10)))
+				bbs.entry.created.at = system.fido_addr_list[0];
+			if(isNaN(parseInt(bbs.entry.updated.at, 10)))
+				bbs.entry.updated.at = system.fido_addr_list[0];
+		}
+		bbs.imported = true;
+		new_list.push(bbs);
+	}
+	return new_list;
 }
 
 const base64_max_line_len = 72;
